chore(style): moves pages css modules to own folder
